Delete item -> delete image file if exists

+++ routes/index.js
@@ -3534,11 +3534,13 @@ router.post('/item/delete', isAuthenticated, function (req, res ) {
         console.log('Articulo eliminado: ', data[0].id);
 
         // borra la imagen anterior
-        var img_path = path.join(__dirname, '..', 'uploads/', data[0].nombre_imagen);
+        if ( data[0].nombre_imagen!== null ) {
+            const img_path = path.join(__dirname, '..', 'uploads/', data[0].nombre_imagen);
 
-        if ( fs.existSync( img_path )) {
-            fs.unlinkSync(img_path);
-            console.log('successfully deleted ' + img_path);
+            if (fs.existSync(img_path)) {
+                fs.unlinkSync(img_path);
+                console.log('successfully deleted ' + img_path);
+            }
         }
